review done

+++ src/components/RecipeCard.jsx
@@ -4,13 +4,21 @@ import { deleteRecipe } from '../services/Auth'
 
 const RecipeCard = ({ recipe, onDelete, user }) => {
   const handleDelete = async () => {
-    try {
-      await deleteRecipe(recipe._id)
-      if (onDelete) {
-        onDelete(recipe._id)
+    const confirmDelete = window.confirm(
+      'Are you sure you want to delete this recipe?'
+    )
+
+    if (confirmDelete) {
+      try {
+        await deleteRecipe(recipe._id) // Call deleteRecipe method
+        if (onDelete) {
+          onDelete(recipe._id) // Call onDelete prop to update parent state
+        }
+        // Reload the page after successful deletion
+        window.location.reload()
+      } catch (error) {
+        console.error('Error deleting recipe:', error)
       }
-    } catch (error) {
-      console.error('Error deleting recipe:', error)
     }
   }
 
@@ -18,7 +26,7 @@ const RecipeCard = ({ recipe, onDelete, user }) => {
 
   return (
     <div className="recipe-card">
-      <Link to={`/recipes/${recipe._id}`} user={user}>
+      <Link to={`/recipes/${recipe._id}`}>
         <h3>{recipe.name}</h3>
       </Link>
       <div className="card-buttons">
